+++ src/clues/__tests__/avoid-emoji.test.ts
