+++ src/clues/avoid-emoji.ts
