+++ src/clues/decorative.ts
