+++ src/clues/end-with.ts
