+++ src/clues/image-link.ts
