+++ src/clues/no-alt.ts
